app/requirements.txt
fastapi
uvicorn
gunicorn
jinja2
python-multipart
gradio
python-dotenv
openai
pandas
azure-core
azure-cosmos
numpy